"final push need a break"
I think if I hadn't gotten stuck on the missing ">" I could have finished this assigment.

I definitely need practice with functions and dom manipulation. I understand how to do it, but it is not second nature...

I need to develop the best method for approaching the problem. more DOM manipulation practice for sure.